+++ app/controllers/lists_controller.rb
@@ -1,10 +1,12 @@
 class ListsController < ApplicationController
   def new
     @list = board.lists.new
+    authorize @list
   end
 
   def create
     @list = board.lists.new(list_params)
+    authorize @list
     if @list.save
       redirect_to root_path, notice: 'List was successfully created.'
     else
@@ -13,11 +15,12 @@ class ListsController < ApplicationController
   end
 
   def edit
-    list
+    authorize list
     board
   end
 
   def update
+    authorize list
     if list.update(list_params)
       redirect_to root_path, notice: 'List was successfully updated.'
     else
@@ -26,6 +29,7 @@ class ListsController < ApplicationController
   end
 
   def destroy
+    authorize list
     list.destroy
     redirect_to root_path, notice: 'List was successfully destroyed.'
   end
